fixed babel config settings to include transform-runtime
