+++ babel.config.json
@@ -3,5 +3,12 @@
         "@babel/preset-env",
         "@babel/preset-react"
     ],
-    "compact": true
+    "compact": true,
+    "env": {
+        "test": {
+            "plugins": [
+                "@babel/plugin-transform-runtime"
+            ]
+        }
+    }
 }
