+++ package.json
@@ -29,6 +29,7 @@
     "xcode": "^3.0.1"
   },
   "devDependencies": {
+    "@babel/plugin-transform-runtime": "^7.10.3",
     "babel-jest": "^26.1.0",
     "babel-polyfill": "^6.26.0",
     "supertest": "^4.0.2",
